Connection of our Backend with the Frontend
In this section we will work integrating our backend with the frontend made in React, which we have almost finished, we just have to call the services and trigger the respective actions that we already have in our calendar application.

This whole section is focused on the part of authentication and handling of the JWT, you have to save and update the authentication token and perform the respective tests!